+++ Questionable/Windows/JournalComponents/QuestRewardComponent.cs
@@ -68,7 +68,7 @@ internal sealed class QuestRewardComponent
                         ? QstTheme.Success
                         : QstTheme.Danger;
                 FontAwesomeIcon icon = complete ? FontAwesomeIcon.Check : FontAwesomeIcon.Times;
-                if (uiUtils.ChecklistItem(name, color, icon, iconOverride: questJournalUtils.GetIconOverride(q, icon)))
+                if (uiUtils.ChecklistItem(name, color, icon, iconOverride: QuestJournalUtils.GetIconOverride(q, icon)))
                 {
                     using ImRaii.TooltipDisposable tooltip = ImRaii.Tooltip();
                     ImGui.Text(_LF("Obtained from: {0}", q.Name));
@@ -107,7 +107,7 @@ internal sealed class QuestRewardComponent
                         ? QstTheme.Success
                         : QstTheme.Danger;
                 FontAwesomeIcon icon = complete ? FontAwesomeIcon.Check : FontAwesomeIcon.Times;
-                if (uiUtils.ChecklistItem(name, color, icon, iconOverride: questJournalUtils.GetIconOverride((QuestInfo)questInfo, icon)))
+                if (uiUtils.ChecklistItem(name, color, icon, iconOverride: QuestJournalUtils.GetIconOverride((QuestInfo)questInfo, icon)))
                 {
                     using ImRaii.TooltipDisposable tooltip = ImRaii.Tooltip();
                     if (item.Type is not EItemRewardType.TripleTriadCard)